Move REST API Classes to Rest namespace
