+++ lib/uploadcare-rails.rb
@@ -2,10 +2,10 @@
 
 require 'uploadcare/rails/engine'
 require 'uploadcare/rails/configuration'
-require 'uploadcare/rails/api/file_api'
-require 'uploadcare/rails/api/group_api'
-require 'uploadcare/rails/api/project_api'
-require 'uploadcare/rails/api/webhook_api'
+require 'uploadcare/rails/api/rest/file_api'
+require 'uploadcare/rails/api/rest/group_api'
+require 'uploadcare/rails/api/rest/project_api'
+require 'uploadcare/rails/api/rest/webhook_api'
 
 module Uploadcare
   # A root module for the gem
